rename jtcCli file
