+++ bin/models/makeResumeCLI.js
